requirements.txt
apache-airflow[celery]==3.1.2
dbt-trino==1.9.3
dbt-core==1.10.10
